Remote detail - split tabs into files
