+++ src/containers/ansible-remote/tab-access.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { AnsibleRemoteType } from 'src/api';
+import { Details } from 'src/components';
+
+interface TabProps {
+  item: AnsibleRemoteType;
+  actionContext: object;
+}
+
+export const AccessTab = ({ item }: TabProps) => <Details item={item} />;
